WIP - Spike integration of Opentracing standard
This is a _very_ rough pass at implementing the `opentracing` gem, which
is created to follow the [Opentracing
Standard](https://opentracing.io/). Tools integrating with opentracing
are supported out of the box by a number of APM tools available in the
industry, and connecting support for others becomes much easier through
the standard interface provided.  This will allow more granular
performance optimization than the log debugger we have in place now.

As I said above, this is still pretty rough and doesn't have specs or
anything, but I wanted to get something out for discussion purposes. I'd
ideally like to use ActiveSupport::Notifications the same way we are
already doign for the debug log, but as far as I know it only supports
point-in-time notifications, as opposed to wrapping the entire timeframe
for a call, which is important for tracing.

Right now the debugger and the tracer are both using slightly different
event names as well (and the things they are tracking don't 100%
overlap yet). Ideally both of these could be resolved before this makes
it into master.